README.md
This repository will be dedicated to showcase a personal portfolio made with the help of react.js. I have used tailwind for layout and js for making the page interactive.
It includes my personal experiences, contact information and form of connection. Feel free to drop a like on some of my repos.